Add support for action lists.
An InhibitorObject can define a dictionary of actions where each action
is associated with a list of functions that must be run.  These are then
wrapped by run_action in InhibitorObject.

Basically this means we can easily drop status files so that costly
operations do not have to be run twice.